check purchasing process

+++ sever/Bee_RushTech/src/main/java/com/project/bee_rushtech/services/OrderDetailService.java
@@ -83,10 +83,13 @@ public class OrderDetailService implements IOrderDetailService{
             }
         }
         // Xử lý số lượng sản phẩm
+        if(product.getQuantity()-cartItem.getQuantity()<0){
+            throw new InvalidException("Product out of stock");
+        }
         product.setQuantity(product.getQuantity()-cartItem.getQuantity());
 
         //xử lý số lượt thuê
-        product.setRentedQuantity(product.getRentedQuantity()+1);
+        product.setRentedQuantity(product.getRentedQuantity()+cartItem.getQuantity());
         productRepository.save(product);
         // Xóa cart item
         cartItemRepository.delete(cartItem);
